cleanup models part 2
- Deprecation
- Relase
- Dependency
- JavaDependency
- ScalaDependency
- Scope
- ScalaTargets
- ISO_8601_Date
