+++ data/src/main/scala/ch.epfl.scala.index.data/project/ProjectConvert.scala
@@ -6,10 +6,9 @@ import cleanup._
 import model._
 import model.misc._
 import bintray._
+import ch.epfl.scala.index.model.release.{Dependency, JavaDependency, ScalaDependency, Scope}
 import github._
-
 import me.tongfei.progressbar._
-
 import org.joda.time.DateTime
 
 object ProjectConvert {
