+++ data/src/test/scala/ch.epfl.scala.index/cleanup/ArtifactNameParserTests.scala
@@ -2,8 +2,8 @@ package ch.epfl.scala.index
 package data
 package cleanup
 
+import ch.epfl.scala.index.model.release.ScalaTargets
 import model._
-
 import utest._
 
 object ArtifactNameParserTest extends TestSuite{
