+++ model/src/main/scala/ch.epfl.scala.index.model/Deprecation.scala
@@ -1,8 +1,11 @@
 package ch.epfl.scala.index.model
 
+import ch.epfl.scala.index.model.misc.ISO_8601_Date
+
 /**
  * Deprecation class to mark an artifact as deprecated and point to a bunch of
  * alternatives instead.
+ *
  * @param since the date since the artifact is deprecated
  * @param useInstead a list of alternate Artifact references
  */
