+++ model/src/main/scala/ch.epfl.scala.index.model/Release.scala
@@ -1,5 +1,8 @@
 package ch.epfl.scala.index.model
 
+import ch.epfl.scala.index.model.misc.{GeneralReference, ISO_8601_Date, MavenReference}
+import ch.epfl.scala.index.model.release._
+
 // typelevel/cats-core (scalajs 0.6, scala 2.11) 0.6.0
 case class Release(
   // famous maven triple: org.typelevel - cats-core_sjs0.6_2.11 - 0.6.0
@@ -16,8 +19,8 @@ case class Release(
   licenses: Set[License] = Set(),
 
   /** split dependencies in 2 fields because elastic can't handle 2 different types
-    * in one field. That is a simple workaround for that
-    */
+   * in one field. That is a simple workaround for that
+   */
   scalaDependencies: Seq[ScalaDependency] = Seq(),
   javaDependencies: Seq[JavaDependency] = Seq(),
   reverseDependencies: Seq[ScalaDependency] = Seq()
@@ -25,14 +28,15 @@ case class Release(
   def sbtInstall = {
     val scalaJs = reference.targets.scalaJsVersion.isDefined
     val crossFull = reference.targets.scalaVersion.patch.isDefined
-  
+
     val (artifactOperator, crossSuffix) =
-      if (scalaJs)       ("%%%",                         "")
-      else if(crossFull) (  "%", " cross CrossVersion.full")
-      else               ( "%%",                         "")
+      if (scalaJs) ("%%%", "")
+      else if (crossFull) ("%", " cross CrossVersion.full")
+      else ("%%", "")
 
     s""""${maven.groupId}" $artifactOperator "${reference.artifact}" % "${reference.version}$crossSuffix""""
   }
+
   def mavenInstall = {
     import maven._
     s"""|<dependency>
@@ -41,12 +45,14 @@ case class Release(
         |  <version>$version</version>
         |</dependency>""".stripMargin
   }
+
   def gradleInstall = {
     import maven._
     s"compile group: '$groupId', name: '$artifactId', version: '$version'"
   }
+
   def scalaDocURI: Option[String] = {
-    if(mavenCentral) {
+    if (mavenCentral) {
       import maven._
       // no frame
       // hosted on s3 at:
@@ -73,27 +79,20 @@ case class Release(
   lazy val dependencyCount = scalaDependencies.size + javaDependencies.size
 
   /**
-    * collect a list of version for a reverse dependency
-    * @param dep current looking dependency
-    * @return
-    */
-  def versionsForReverseDependencies(dep: ScalaDependency): Seq[SemanticVersion] =  {
+   * collect a list of version for a reverse dependency
+   *
+   * @param dep current looking dependency
+   * @return
+   */
+  def versionsForReverseDependencies(dep: ScalaDependency): Seq[SemanticVersion] = {
 
     reverseDependencies.filter(d => d.dependency.name == dep.dependency.name).map(_.dependency.version)
   }
 }
 
-/**
-  * General Reference to Group MavenReference and Release.Reference
-  * to a category form simpler usage.
-  */
-sealed trait GeneralReference {
 
-  def name: String
-  def httpUrl: String
-}
+object Release {
 
-object Release{
   case class Reference(
     organization: String, // typelevel               | akka
     artifact: String, // cats-core               | akka-http-experimental
@@ -102,91 +101,8 @@ object Release{
   ) extends GeneralReference {
 
     def name: String = s"$organization/$artifact"
+
     def httpUrl: String = s"/$organization/$artifact/$version"
   }
+
 }
-
-// com.typesafe.akka - akka-http-experimental_2.11 - 2.4.6 | org.typelevel - cats-core_sjs0.6_2.11 - 0.6.0
-case class MavenReference(
-   groupId: String, // org.typelevel         | com.typesafe.akka
-   artifactId: String, // cats-core_sjs0.6_2.11 | akka-http-experimental_2.11
-   version: String // 0.6.0                 | 2.4.6
-) extends GeneralReference {
-
-  def name: String = s"$groupId/$artifactId"
-  def httpUrl: String = s"http://search.maven.org/#artifactdetails|$groupId|$artifactId|$version|jar"
-}
-
-case class ScalaTargets(scalaVersion: SemanticVersion, scalaJsVersion: Option[SemanticVersion] = None) {
-
-  /** simple modifier for display a nice name */
-  lazy val name: String = scalaJsVersion.map(v => s"Scala.js ${v.toString} ($scalaVersion)").getOrElse(s"Scala $scalaVersion")
-
-  /** converting the scala version for support tags in elastic  - only major.minor to have small list */
-  lazy val supportName: String = scalaJsVersion.map(v => s"scala.js_${v.major}.${v.minor}").getOrElse(s"scala_${scalaVersion.major}.${scalaVersion.minor}")
-
-  /** simple modifier for ordering */
-  lazy val orderName: String = scalaJsVersion.map(v => s"${scalaVersion.toString.replace(".", "")}_${v.toString.replace(".", "")}").getOrElse(scalaVersion.toString.replace(".", ""))
-}
-
-case class ISO_8601_Date(value: String) // 2016-05-20T12:48:52.533-04:00
-
-sealed trait Scope {
-  def name: String
-}
-object Scope {
-
-  /** Sbt Scopes: http://www.scala-sbt.org/0.12.2/docs/Getting-Started/Scopes.html
-    * Maven Scopes: http://maven.apache.org/guides/introduction/introduction-to-dependency-mechanism.html#Dependency_Scope
-    */
-  case object Test extends Scope {
-    val name = "test"
-  }
-  case object Provided extends Scope {
-    val name = "provided"
-  }
-  case object Compile extends Scope {
-    val name = "compile"
-  }
-  case object Runtime extends Scope {
-    val name = "runtime"
-  }
-  case object Optional extends Scope {
-    val name = "optional"
-  }
-  case object System extends Scope {
-    val name = "system"
-  }
-  case object Import extends Scope {
-    val name = "import"
-  }
-
-  /**
-    * Convert String scope to Scope class
- *
-    * @param scope the current scope
-    * @return
-    */
-  def apply(scope: String): Scope = scope.toLowerCase match {
-
-    case Test.name => Test
-    case Provided.name => Provided
-    case Compile.name => Compile
-    case Runtime.name => Runtime
-    case Optional.name => Optional
-    case System.name => System
-    case Import.name => Import
-
-    case unknown =>
-      throw new RuntimeException(s"unknown scope $unknown detected during pom parsing.")
-  }
-}
-
-sealed trait Dependency {
-
-  def dependency: GeneralReference
-  def scope: Option[Scope]
-}
-
-case class ScalaDependency(dependency: Release.Reference, scope: Option[Scope]) extends Dependency
-case class JavaDependency(dependency: MavenReference, scope: Option[Scope]) extends Dependency
