+++ model/src/main/scala/ch.epfl.scala.index.model/misc/GeneralReference.scala
@@ -1,0 +1,11 @@
+package ch.epfl.scala.index.model.misc
+
+/**
+ * General Reference to Group MavenReference and Release.Reference
+ * to a category form simpler usage.
+ */
+trait GeneralReference {
+
+  def name: String
+  def httpUrl: String
+}
