+++ model/src/main/scala/ch.epfl.scala.index.model/misc/ISO_8601_Date.scala
@@ -1,0 +1,7 @@
+package ch.epfl.scala.index.model.misc
+
+/**
+ * Iso date - 8601
+ * @param value
+ */
+case class ISO_8601_Date(value: String)
