+++ model/src/main/scala/ch.epfl.scala.index.model/release/Dependency.scala
@@ -1,0 +1,12 @@
+package ch.epfl.scala.index.model.release
+
+import ch.epfl.scala.index.model.misc.GeneralReference
+
+/**
+ * Dependency trait to mark a class as a dependency
+ */
+trait Dependency {
+
+  val dependency: GeneralReference
+  val scope: Option[Scope]
+}
